+++ src/traces/isosurface/attributes.js
@@ -11,7 +11,6 @@
 var colorscaleAttrs = require('../../components/colorscale/attributes');
 var colorbarAttrs = require('../../components/colorbar/attributes');
 var hovertemplateAttrs = require('../../components/fx/hovertemplate_attributes');
-var surfaceAtts = require('../surface/attributes');
 var meshAttrs = require('../mesh3d/attributes');
 var baseAttrs = require('../../plots/attributes');
 
@@ -238,37 +237,16 @@ colorscaleAttrs('', {
 
     colorbar: colorbarAttrs,
 
-    // Flat shaded mode
-    flatshading: {
-        valType: 'boolean',
-        role: 'style',
-        dflt: false,
-        description: [
-            'Determines whether or not normal smoothing is applied to the isosurfaces,',
-            'creating isosurfaces with an angular, low-poly look via flat reflections.'
-        ].join(' ')
-    },
-
-    contour: {
-        show: extendFlat({}, surfaceAtts.contours.x.show, {
-            description: [
-                'Sets whether or not dynamic contours are shown on hover.',
-                'Contours are more useful when hovering on caps and slices.'
-            ].join(' ')
-        }),
-        color: surfaceAtts.contours.x.color,
-        width: surfaceAtts.contours.x.width
-    },
-
-    lightposition: {
-        x: extendFlat({}, surfaceAtts.lightposition.x, {dflt: 1e5}),
-        y: extendFlat({}, surfaceAtts.lightposition.y, {dflt: 1e5}),
-        z: extendFlat({}, surfaceAtts.lightposition.z, {dflt: 0})
-    },
+    contour: meshAttrs.contour,
+    lightposition: meshAttrs.lightposition,
     lighting: meshAttrs.lighting,
+    flatshading: meshAttrs.flatshading,
 
     hoverinfo: extendFlat({}, baseAttrs.hoverinfo)
 }), 'calc', 'nested');
+
+// required defaults to speed up surface normal calculations
+attrs.flatshading.dflt = true; attrs.lighting.facenormalsepsilon.dflt = 0;
 
 attrs.x.editType = attrs.y.editType = attrs.z.editType = attrs.value.editType = 'calc+clearAxisTypes';
 attrs.transforms = undefined;