+++ src/traces/sankey/plot.js
@@ -94,7 +94,6 @@ function linkNonHoveredStyle(d, sankey, visitNodes, sankeyLink) {
     var label = sankeyLink.datum().link.label;
 
     sankeyLink.style('fill-opacity', function(d) {return d.tinyColorAlpha;});
-
     if(label) {
         ownTrace(sankey, d)
             .selectAll('.' + cn.sankeyLink)
@@ -152,16 +151,23 @@ module.exports = function plot(gd, calcData) {
         var obj = d.link.trace.link;
         if(obj.hoverinfo === 'none' || obj.hoverinfo === 'skip') return;
         var rootBBox = gd._fullLayout._paperdiv.node().getBoundingClientRect();
-        var boundingBox = element.getBoundingClientRect();
-        var hoverCenterX = boundingBox.left + boundingBox.width / 2;
-        var hoverCenterY = boundingBox.top + boundingBox.height / 2;
+        var hoverCenterX;
+        var hoverCenterY;
+        if(d.link.circular) {
+            hoverCenterX = (d.link.circularPathData.leftInnerExtent + d.link.circularPathData.rightInnerExtent) / 2 + d.parent.translateX;
+            hoverCenterY = d.link.circularPathData.verticalFullExtent + d.parent.translateY;
+        } else {
+            var boundingBox = element.getBoundingClientRect();
+            hoverCenterX = boundingBox.left + boundingBox.width / 2 - rootBBox.left;
+            hoverCenterY = boundingBox.top + boundingBox.height / 2 - rootBBox.top;
+        }
 
         var hovertemplateLabels = {valueLabel: d3.format(d.valueFormat)(d.link.value) + d.valueSuffix};
         d.link.fullData = d.link.trace;
 
         var tooltip = Fx.loneHover({
-            x: hoverCenterX - rootBBox.left,
-            y: hoverCenterY - rootBBox.top,
+            x: hoverCenterX,
+            y: hoverCenterY,
             name: hovertemplateLabels.valueLabel,
             text: [
                 d.link.label || '',